Move config and credentials providers to separate files
